+++ CHANGELOG.md
@@ -75,6 +75,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - query coverage, density + position based ranking (#159)
 - Link server to mithril managers
 - add order sensitive ranking & BM25 (#166)
+- performance improvements to position index (#172)
 
 ### Fixed